+++ cmd/aar/main.go
@@ -4,10 +4,8 @@ import (
 	"flag"
 	"fmt"
 	"os"
-	"syscall"
 
 	"github.com/angelsolaorbaiceta/aar/cmd"
-	"golang.org/x/term"
 )
 
 func main() {
@@ -59,14 +57,14 @@ func main() {
 	case "encrypt":
 		encryptCmd.Parse(os.Args[2:])
 		validateFileName(*encryptFileNameFlag)
-		password := promptPassword()
+		password := cmd.PromptPassword()
 
 		cmd.EncryptArchive(*encryptFileNameFlag, password)
 
 	case "decrypt":
 		decryptCmd.Parse(os.Args[2:])
 		validateFileName(*decryptFileNameFlag)
-		password := promptPassword()
+		password := cmd.PromptPassword()
 
 		cmd.DecryptArchive(*decryptFileNameFlag, password)
 
@@ -90,29 +88,4 @@ func createArchive(fileName string, fileNames []string) {
 	}
 
 	cmd.CreateArchive(fileName, fileNames)
-}
-
-func promptPassword() string {
-	fmt.Print("Password: ")
-
-	// Disable input echoing
-	passwordBytes, err := term.ReadPassword(int(syscall.Stdin))
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error reading password: %v\n", err)
-		os.Exit(1)
-	}
-
-	fmt.Println() // Move to the next line after password input
-
-	password := string(passwordBytes)
-	validatePassword(password)
-
-	return password
-}
-
-func validatePassword(password string) {
-	if len(password) < 8 {
-		fmt.Fprintf(os.Stderr, "The password must be at least 8 characters long.\n")
-		os.Exit(1)
-	}
 }
